[alpha_factory] refactor orchestrator into modules
